refactor(Cards) : 메서드 정리
